% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedWrapper.R
\name{adlassoMixWrapper}
\alias{adlassoMixWrapper}
\title{New wrapper functions for each method, they source the function script highmed2019.r}
\usage{
adlassoMixWrapper(
  Z,
  X,
  X0,
  y,
  kernel = "linear",
  ncores = 1,
  pmax = length(y) - 2
)
}
\description{
New wrapper functions for each method, they source the function script highmed2019.r
}
