% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSKAT.R
\name{Get_Liu_Params_Mod_Lambda}
\alias{Get_Liu_Params_Mod_Lambda}
\title{functions from SKAT package, with minor modification}
\usage{
Get_Liu_Params_Mod_Lambda(lambda)
}
\description{
functions from SKAT package, with minor modification
}
