% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getpca}
\alias{getpca}
\title{\code{Obtain the first several PCs}}
\usage{
getpca(Z, p = 3)
}
\arguments{
\item{Z}{The `n x m` genotype matrix of data. [matrix, ]}

\item{p}{The first n number of PCs. [num, default p=3]}
}
\value{
return a batch of shell scripts.
}
\description{
\code{Obtain the first several PCs}
}
\examples{
Z <- fread("input/Z_matrix.txt", header=T,data.table=FALSE)
Z = as.matrix(Z)
X0 <- getpca(Z, p=3)

}
