% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.R
\name{reportDirectSNPs}
\alias{reportDirectSNPs}
\title{a function extracting the mediators making the cut}
\usage{
reportDirectSNPs(mod = mod.fixed, tst = tests.fixed)
}
\description{
a function extracting the mediators making the cut
}
