% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_wrapper.R
\name{run_gma}
\alias{run_gma}
\title{\code{Run Genome-wide mediation analysis (GMA)}}
\usage{
run_gma(
  ti = 2,
  cutoff_pm = 0.05,
  ncores = 12,
  phenofile = "data/geno_trait.txt",
 
    genofile = "largedata/geno/allchr_bisnp_n282_snpid_maf01_geno2_pruned_NA_0_matrix.txt",
  rnafile = "largedata/RNA-seq/filtered_GRoot.csv",
  pcfile = "largedata/allchr_bisnp_n282_snpid_maf01_geno2_pruned.eigenvec",
  outdir = "largedata/med_output"
)
}
\arguments{
\item{outdir}{The dir of shell files.}

\item{inputdf}{An input data.frame, with columns of file and out.}

\item{rcodes}{The abosulte path of your R codes to run.}

\item{arrayshid}{The sbatch id.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{cmdno}{Number of commands per CPU, i.e. number of rows per inputdf.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
\code{Run Genome-wide mediation analysis (GMA)}
}
\examples{
run_Rcodes(inputdf=data.frame(file=1:11, out=10), outdir="slurm-script", cmdno=10,
           rcodes = "lib/C_format.R", arrayshid = "slurm-script/run_rcode_array.sh",
           email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}
